Add files via upload
Homes of the 21st century will become more and more selfcontrolled and automated due to the comfort it provides, especially when employed in a private home. A home automation system is a means that allow users to control electric appliances of varying kind. Many existing, well-established home automation systems are based on wired communication. This does not pose a problem until the system is planned well in advance and installed during the physical construction of the building. But for already existing buildings the implementation cost goes very high. In contrast, Wireless systems can be of great help for automation systems. With the advancement of wireless technologies such as Wi-Fi, cloud networks in the recent past, wireless systems are used every day and everywhere.
This paper proposes a Home Automation system that employs the integration of multi-touch mobile devices, cloud networking, wireless communication, and power-line communication to provide the user with remote control of various lights and appliances within their home. This system uses a consolidation of a mobile phone application, handheld wireless remote and PC based program to provide a means of user interface to the consumer.
Home automation systems face four main challenges, these are high cost of ownership, inflexibility, poor manageability, and difficulty in achieving security. The main objectives of this research is to design and implement a home automation system using IoT that is capable of controlling and automating most of the house appliances through an easy manageable web interface. The proposed system has a great flexibility by using Wi-Fi technology to interconnect its distributed sensors to home automation server. This will decrease the deployment cost and will increase the ability of upgrading, and system reconfiguration
The proposed system is a distributed home automation system, consists of server, sensors. Server controls and monitors the various sensors, and can be easily configured to handle more hardware interface module (sensors). The Intel Galileo development board, with built in WiFi card port to which the card is inserted, acts as web server. Automation System can be accessed from the web browser of any local PC in the same LAN using server IP or remotely from any PC or mobile handheld device connected to the internet with appropriate web browser through server real IP (internet IP). WiFi technology is selected to be the network infrastructure that connects server and the sensors. WiFi is chosen to improve system security (by using secure WiFi connection), and to increase system mobility and scalability.